requirements.txt
anyio==3.6.1
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
autopep8==1.7.0
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
dnspython==2.2.1
email-validator==1.2.1
entrypoints==0.4
executing==1.0.0
fastapi==0.81.0
flake8==5.0.4
fonttools==4.37.1
greenlet==1.1.3
h11==0.13.0
httptools==0.4.0
idna==3.3
ipykernel==6.15.2
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
nest-asyncio==1.5.5
numpy==1.23.2
orjson==3.8.0
packaging==21.3
pandas==1.4.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pydantic==1.10.1
pyflakes==2.5.0
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.2.1
requests==2.28.1
seaborn==0.12.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
stack-data==0.5.0
starlette==0.19.1
toml==0.10.2
tornado==6.2
traitlets==5.3.0
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.12
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.3
